adjusted ClusterOverview
