+++ public/i18n/en.yaml
@@ -5,6 +5,8 @@ apps:
   title: Apps
 cluster-overview:
   headers:
+    resource-details: Resource Details
+    monitoring-and-health: Monitoring and Health
     cpu: CPU
     memory: Memory
     metrics: Metrics
@@ -13,8 +15,12 @@ cluster-overview:
     cpu-usage-m: 'CPU Usage'
     memory-usage-gib: 'Memory Usage'
   tooltips:
-    cpu-used: 'CPU (m) used: {{value}}/{{max}}'
-    memory-used: 'Memory (GiB) used: {{value}}/{{max}}'
+    percentage:
+      cpu-used: 'CPU used: {{percentage}}'
+      memory-used: 'Memory used: {{percentage}}'
+    units:
+      cpu-used: 'CPU (m) used: {{value}}/{{max}}'
+      memory-used: 'Memory (GiB) used: {{value}}/{{max}}'
 cluster-role-bindings:
   description: <0>ClusterRoleBinding</0> assigns the permissions defined in a ClusterRole to a user.
   name_singular: ClusterRoleBinding
