+++ src/components/Clusters/views/ClusterOverview/ClusterOverview.js
@@ -3,9 +3,10 @@ import React from 'react';
 import { ClusterOverviewHeader } from './ClusterOverviewHeader';
 import { ClusterNodes } from './ClusterNodes/ClusterNodes';
 
-import './ClusterOverview.scss';
 import { ClusterValidation } from './ClusterValidation/ClusterValidation';
 import { useFeature } from 'hooks/useFeature';
+import { useNodesQuery } from 'components/Nodes/nodeQueries';
+import ClusterStats from './ClusterStats';
 
 const Injections = React.lazy(() =>
   import('../../../Extensibility/ExtensibilityInjections'),
@@ -13,6 +14,8 @@ const Injections = React.lazy(() =>
 
 export function ClusterOverview() {
   const clusterValidation = useFeature('CLUSTER_VALIDATION');
+  const { nodes: data, error, loading } = useNodesQuery();
+
   return (
     <ClusterOverviewHeader
       content={
@@ -22,7 +25,8 @@ export function ClusterOverview() {
             slot="details-top"
             root=""
           />
-          <ClusterNodes />
+          {data && <ClusterStats data={data} />}
+          <ClusterNodes data={data} error={error} loading={loading} />
           {clusterValidation?.isEnabled && <ClusterValidation />}
           <Injections destination="ClusterOverview" slot="details-bottom" />
         </>
