--- src/components/Clusters/views/ClusterOverview/ClusterOverview.scss
@@ -1,3 +1,0 @@
-.gardener-provider {
-  text-transform: uppercase;
-}
